+++ src/internal/path.rs
@@ -93,7 +93,7 @@ impl WhichWay {
     /// Given a height and an index of a leaf, determine which direction the path down to that leaf
     /// should branch at the node at that height.
     #[inline]
-    pub fn at(height: u64, index: u64) -> WhichWay {
+    pub fn at(height: u8, index: u64) -> WhichWay {
         // Shift the index right by (2 * (height - 1)) so that the last 2 bits are our direction, then
         // mask off just those bits and branch on them to generate the output
         match (index >> (2 * (height - 1))) & 0b11 {
@@ -112,7 +112,7 @@ mod test {
     use proptest::prelude::*;
 
     /// Get directions from the root (at the given height)
-    fn directions_of_index(height: u64, index: u64) -> Vec<WhichWay> {
+    fn directions_of_index(height: u8, index: u64) -> Vec<WhichWay> {
         (1..=height + 1)
             .rev() // iterate from the root to the leaf (height down to 1)
             .map(|height| WhichWay::at(height, index))
@@ -144,7 +144,7 @@ mod test {
         fn which_way_correct(
         (height, index) in (
             // This is a dependent generator: we ensure that the index is in-bounds for the height
-            (0u64..(3 * 8)), 0u64..u64::MAX).prop_map(|(height, index)| (height, (index % (4u64.pow(height as u32)))))
+            (0u8..(3 * 8)), 0u64..u64::MAX).prop_map(|(height, index)| (height, (index % (4u64.pow(height as u32)))))
         ) {
             assert_eq!(index, index_of_directions(&directions_of_index(height, index)));
         }
